MedHive-main/MedHive-main/backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
sqlalchemy==2.0.22
psycopg2-binary==2.9.9
alembic==1.12.0
python-jose==3.3.0
passlib==1.7.4
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6
email-validator==2.1.0
bcrypt==4.0.1
python-dotenv==1.0.0
langchain==0.0.313
groq==0.3.0
astra-db-python==0.10.0
mlflow==2.8.0
flwr==1.5.0
huggingface-hub==0.19.4
scikit-learn==1.3.2
gunicorn==21.2.0
pytest==7.4.3 